ALLJOYN-2210 Fixed compiler warnings from Windows VC++
ICECandidatePair was reporting a warning C4244 'return' conversion from double
to uint32_t possible loss of data. The double was being returned from the
GetQueuedTimeOffset() method from StunRetry.

I talked with Padama and she verified that it was safe to change the return
value of the GetQueuedTimeOffset function.

Fixed a few implicit other casting warnings that could result in possible loss of data.

+++ daemon/ice/StunRetry.cc
@@ -5,7 +5,7 @@
  */
 
 /******************************************************************************
- * Copyright 2009,2012 Qualcomm Innovation Center, Inc.
+ * Copyright 2009,2012-2013 Qualcomm Innovation Center, Inc.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -45,7 +45,7 @@ bool CheckRetry::RetryAvailable(void)
     return (AnyRetriesNotSent() && RetryTimedOut());
 }
 
-double CheckRetry::GetQueuedTimeOffset(void)
+uint32_t CheckRetry::GetQueuedTimeOffset(void)
 {
     return (queuedTime + maxReceiveWaitMsec[sendAttempt]);
 }
